requirement.txt
Flask
Flask-Login
pymongo
pandas
pywebview
Werkzeug
pyinstaller
waitress
